Support 1D conv (32 bit only)

+++ common/c_callbacks.h
@@ -7,7 +7,8 @@
 extern "C" {
 #endif
 struct ParameterInfo;
-typedef void (*data_preservation_func)(struct ParameterInfo *param, uint16_t offset_in_word, const void *src, size_t n, uint16_t timer_delay);
+// FIXME: descript in common/platform.cpp
+typedef void (*data_preservation_func)(struct ParameterInfo *param, uint32_t offset_in_word, const void *src, size_t n, uint16_t timer_delay);
 void record_overflow_handling_overhead(uint32_t cycles);
 #ifdef __cplusplus
 }
